RUMM-2029 Refactor the way `RUMMonitor` is intizialized from `RUMFeature`
to isolate `RUMDependencies` creation into separate method. This way we can
partially replace some dependencies in certain tests.
